Implemented alert dialog instead of a new activity when deleting a book
