+++ app/src/main/res/values/styles.xml
@@ -34,4 +34,8 @@
         <item name="android:windowIsTranslucent">true</item>
         <item name="android:windowCloseOnTouchOutside">true</item>
     </style>
+
+    <style name="AppTheme.AlertDialog">
+        <item name="android:windowCloseOnTouchOutside">true</item>
+    </style>
 </resources>
